Add Classes and IDs to spaghetti.html
Added all the necessary classes to the spaghetti.html file for the
necessary styles to be applied across the page. Ran into a few issues
with certain styles not being applied and had to create new CSS and
selectors to rectify the issue. Ensured that the changes made to CSS
did not affect the other hmtl files.